% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_functions.R
\name{BCORSIS}
\alias{BCORSIS}
\title{Ball Correlation Sure Independence Screening (BCORSIS)}
\usage{
BCORSIS(df, time, status, features, d = "large")
}
\arguments{
\item{df}{data frame including columns with features and time-to-event endpoint information}

\item{time}{character/column name containing survival time}

\item{status}{character/column name containing survival status}

\item{features}{vector of column names or column indices containing features}

\item{d}{see \code{bcorsis}}
}
\value{
a vector with Ball correlation for each feature
}
\description{
Ball Correlation Sure Independence Screening (BCORSIS)
}
\details{
only a wrapper for \code{bcorsis} function in the \code{Ball} R package
}
\examples{
\dontrun{
require("survival")
require("penalized")
require("Ball")
data(nki70)
BCORSIS(nki70,"time","event", colnames(nki70)[-c(1:7)])[1:10]
}
}
\references{
Pan, W., Wang, X., Xiao, W., and Zhu, H. (2018). A generic sure independence screening procedure. Journal of the American Statistical Association, to appear.
}
\author{
.
}
