% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcov.R
\name{centmat}
\alias{centmat}
\title{Calculate a centralized version of the distance matrix.}
\usage{
centmat(X, metr.X = "euclidean", type.X = "sample", bias.corr = TRUE, n,
  p, ...)
}
\arguments{
\item{X}{a numeric vector or a numeric matrix.}

\item{metr.X}{metric that should be used to compute the distance matrix.}

\item{type.X}{either "sample" or "distance". If type.X = "sample", X must be}

\item{bias.corr}{logical; indicates if the corresponding estimator of the covariance matrix is biased or unbiased.}

\item{n}{number of samples, i.e. the number of rows of X..}

\item{p}{number of repetitions, i.e. the number of columns of X.}

\item{...}{additional parameters that are used for other metrics (e.g., the bandwidth for Gaussian kernels)}
}
\value{
The centralized distance matrix corresponding to X.
}
\description{
Calculate a centralized version of the distance matrix.
}
\details{
For metr.X the following metrices are built in: euclidean, gaussian and discrete. However,
it is possible to use a function taking two numerical arguments as metr.X.
}
