% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_functions.R
\name{COX}
\alias{COX}
\title{SIS-Cox}
\usage{
COX(df, time, status, features)
}
\arguments{
\item{df}{data frame including columns with features and time-to-event endpoint information}

\item{time}{character/column name containing survival time}

\item{status}{character/column name containing survival status}

\item{features}{vector of column names or column indices containing features}
}
\value{
a named vector with partial log-likelihood for each feature
}
\description{
SIS-Cox
}
\details{
wrapper for \code{coxph} function
}
\examples{
\dontrun{
require("survival")
require("penalized")
data(nki70)
COX(nki70,"time","event", colnames(nki70)[-c(1:7)])[1:10]
}
}
\references{
Fan, J., Feng, Y., and Wu, Y. (2010). High-dimensional variable selection for Cox’s proportional hazards model. In Borrowing Strength: Theory Powering Applications–A Festschrift for Lawrence D. Brown, pages 70-86. Institute of Mathematical Statistics.
}
\author{
.
}
