% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_functions.R
\name{CoxConc}
\alias{CoxConc}
\title{C-Index for each feature}
\usage{
CoxConc(df, time, status, features)
}
\arguments{
\item{df}{data frame including columns with features and time-to-event endpoint information}

\item{time}{character/column name containing survival time}

\item{status}{character/column name containing survival status}

\item{features}{vector of column names or column indices containing features}
}
\value{
a named vector with c-index for each feature
}
\description{
C-Index for each feature
}
\details{
only a wrapper for \code{coxph} function
}
\examples{
\dontrun{
require("survival")
require("penalized")
data(nki70)
CoxConc(nki70,"time","event", colnames(nki70)[-c(1:7)])[1:10]
}
}
\author{
.
}
