% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_functions.R
\name{CRCDCS}
\alias{CRCDCS}
\title{Composite Robust Censored Distance Correlation Screening (CRCDCS)}
\usage{
CRCDCS(df, time, status, features, q = seq(0.1, 0.9, 0.05))
}
\arguments{
\item{df}{data frame including columns with features and time-to-event endpoint information}

\item{time}{character/column name containing survival time}

\item{status}{character/column name containing survival status}

\item{features}{vector of column names or column indices containing features}

\item{q}{vector of quantiles of survival function to be evaluated, default is \code{seq(0.1,0.9,0.05)}}
}
\value{
a named vector with squared distance correlation for each feature
}
\description{
Composite Robust Censored Distance Correlation Screening (CRCDCS)
}
\details{
.
}
\examples{
\dontrun{
require("survival")
require("Rfast")
require("penalized")
data(nki70)
CRCDCS(nki70,"time","event", colnames(nki70)[-c(1:7)])[1:10]
}
}
\references{
Chen, X., Chen, X., and Wang, H. (2018). Robust feature screening for ultra-high dimensional right censored data via distance correlation. Computational Statistics & Data Analysis, 119, 118-138.
}
\author{
Thomas Hielscher
}
