% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_functions.R
\name{CRIS}
\alias{CRIS}
\title{Censored rank independence screening (CRIS)}
\usage{
CRIS(df, time, status, features)
}
\arguments{
\item{df}{data frame including columns with features and time-to-event endpoint information}

\item{time}{character/column name containing survival time}

\item{status}{character/column name containing survival status}

\item{features}{vector of column names or column indices containing features}
}
\value{
a named vector
}
\description{
Censored rank independence screening (CRIS)
}
\details{
Code downloaded from \url{https://rsong.wordpress.ncsu.edu/files/2018/09/cris.zip}
}
\examples{
\dontrun{
require("survival")
require("penalized")
data(nki70)
CRIS(nki70,"time","event", colnames(nki70)[-c(1:7)])[1:10]
}
}
\references{
Song, R., Lu, W., Ma, S., and Jessie Jeng, X. (2014). Censored rank independence screening for high-dimensional survival data. Biometrika, 101, 799-814.
}
\author{
.
}
