% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcov.R
\name{distcorr}
\alias{distcorr}
\title{Calculates the distance correlation}
\usage{
distcorr(X, Y, affine = FALSE, bias.corr = TRUE, type.X = "sample",
  type.Y = "sample", metr.X = "euclidean", metr.Y = "euclidean",
  use = "all")
}
\arguments{
\item{X}{contains either the first sample or its corresponding distance matrix. In the first case, this input can be either a vector of positive length, a matrix with one column or a data.frame with one column. In this case, type.X must be specified as "sample". In the second case, the input must be a distance matrix corresponding to the sample of interest. In this second case, type.X must be "distance".}

\item{Y}{see X.}

\item{affine}{logical; indicates if the affinely transformed distance correlation should be calculated or not.}

\item{bias.corr}{logical; indicates if the bias corrected version of the sample distance correlation should be calculated.}

\item{type.X}{either "sample" or "distance"; specifies the type of input for X.}

\item{type.Y}{see type.X.}

\item{metr.X}{specifies the metric which should be used for X to analyse the distance correlation. TO DO: Provide details for this.}

\item{metr.Y}{see metr.X.}

\item{use}{: "all" uses all observations, "complete.obs" excludes NA's}
}
\value{
numeric giving the distance correlation between samples X and Y.
}
\description{
Calculates the distance correlation
}
