% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcov.R
\name{distmat}
\alias{distmat}
\title{Calculate the distance matrix of a given vector and a given metric.}
\usage{
distmat(X, metr.X = "euclidean", n, p, ...)
}
\arguments{
\item{X}{a numeric vector or a numeric matrix.}

\item{metr.X}{metric that should be used to compute the distance matrix.}

\item{n}{number of samples, i.e. the number of rows of X..}

\item{p}{number of repetitions, i.e. the number of columns of X.}

\item{...}{additional parameters that are used for other metrics (e.g., the bandwidth for Gaussian kernels)}
}
\value{
The distance matrix corresponding to X.
}
\description{
Calculate the distance matrix of a given vector and a given metric.
}
\details{
For metr.X the following metrices are built in: euclidean, gaussian and discrete. However,
it is possible to use a function taking two numerical arguments as metr.X.
}
