% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcov.R
\name{distvar}
\alias{distvar}
\title{Calculates the distance variance}
\usage{
distvar(X, affine = FALSE, bias.corr = TRUE, type.X = "sample",
  metr.X = "euclidean", use = "all")
}
\arguments{
\item{X}{contains either the first sample or its corresponding distance matrix.

In the first case, this input can be either a vector of positive length,

a matrix with one column or a data.frame with one column.

In this case, type.X must be specified as "sample".

In the second case, the input must be a distance matrix corresponding to the sample of interest.

In this second case, type.X must be "distance".}

\item{affine}{logical; indicates if the affinely transformed distance variance should be calculated or not.}

\item{bias.corr}{logical; indicates if the bias corrected version of the sample distance variance should be calculated.}

\item{type.X}{either "sample" or "distance"; specifies the type of input for X.}

\item{metr.X}{specifies the metric which should be used for X to analyse the distance variance TO DO: Provide details for this.}

\item{use}{: "all" uses all observations, "complete.obs" excludes NA's}
}
\value{
numeric giving the distance variance of the sample X..
}
\description{
Calculates the distance variance
}
