% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcov.R
\name{extract_np}
\alias{extract_np}
\title{Extract the dimensions of X.}
\usage{
extract_np(X, type.X)
}
\arguments{
\item{X}{a numeric vector or a numeric matrix.}

\item{type.X}{either "sample" or "distance". If type.X = "sample", X must be
a numeric vector or numeric matrix with the corresponding observations. If metr.X = "distance",
X must be a distance matrix.}
}
\value{
The centralized distance matrix corresponding to X.
}
\description{
Extract the dimensions of X.
}
