% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_functions.R
\name{FAST}
\alias{FAST}
\title{FAST (feature aberration at survival times)}
\usage{
FAST(df, time, status, features)
}
\arguments{
\item{df}{data frame including columns with features and time-to-event endpoint information}

\item{time}{character/column name containing survival time}

\item{status}{character/column name containing survival status}

\item{features}{vector of column names or column indices containing features}
}
\value{
a named vector with absolute coefficient for each feature
}
\description{
FAST (feature aberration at survival times)
}
\details{
only a wrapper for \code{ahaz} function in \code{ahaz} package
}
\examples{
\dontrun{
require("survival")
require("penalized")
require("ahaz")
data(nki70)
FAST(nki70,"time","event", colnames(nki70)[-c(1:7)])[1:10]
}
}
\references{
Gorst-Rasmussen, A. and Scheike, T. (2013). Independent screening for single-index hazard rate models with ultrahigh dimensional features. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 75, 217-245.
}
\author{
.
}
