% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_functions.R
\name{FDR}
\alias{FDR}
\title{False Discovery Rate (FDR) for given model size(s)}
\usage{
FDR(scores, m, trueVars)
}
\arguments{
\item{scores}{scores matrix; output of screening methods for simulated data, larger value means better ranking, larger value means better ranking}

\item{m}{scalar or vector of model sizes}

\item{trueVars}{vector of indices of true variables}
}
\value{
scalar/vector with FDR for each model size
}
\description{
False Discovery Rate (FDR) for given model size(s)
}
\details{
ties are handled using the 'random' approach in \code{rank}.
}
\author{
Manuela Hummel
}
