% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_functions.R
\name{FP}
\alias{FP}
\title{MFP based SIS-Cox}
\usage{
FP(df, time, status, features, fp = 4)
}
\arguments{
\item{df}{data frame including columns with features and time-to-event endpoint information}

\item{time}{character/column name containing survival time}

\item{status}{character/column name containing survival status}

\item{features}{vector of column names or column indices containing features}

\item{fp}{degrees of freedom of the FP model. Default is 4.}
}
\value{
a named vector with partial log-likelihood for each feature
}
\description{
MFP based SIS-Cox
}
\details{
wrapper for \code{mfp} function
}
\examples{
\dontrun{
require("survival")
require("penalized")
require("mfp")
data(nki70)
FP(nki70,"time","event", colnames(nki70)[-c(1:7)])[1:10]
}
}
\references{
Ambler, G. and Benner, A. (2015). mfp: Multivariable Fractional Polynomials. R package version 1.5.2.
}
\author{
.
}
