% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_functions.R
\name{IPOD}
\alias{IPOD}
\title{Integrated powered density (IPOD)}
\usage{
IPOD(df, time, status, features, gamma = 1)
}
\arguments{
\item{df}{data frame including columns with features and time-to-event endpoint information}

\item{time}{character/column name containing survival time}

\item{status}{character/column name containing survival status}

\item{features}{vector of column names or column indices containing features}

\item{gamma}{power of interest, see \code{IPOD.cont}}
}
\value{
a named vector
}
\description{
Integrated powered density (IPOD)
}
\details{
only a wrapper for \code{IPOD.cont} function, which is downloaded from \url{https://github.com/younghhk/software/blob/master/R/IPOD.R}
}
\examples{
\dontrun{
require("survival")
require("penalized")
require("survPresmooth")
data(nki70)
IPOD(nki70,"time","event", colnames(nki70)[-c(1:7)])[1:10]
}
}
\references{
Hong, H. G., Chen, X., Christiani, D. C., and Li, Y. (2018a). Integrated powered density: Screening ultrahigh dimensional covariates with survival outcomes. Biometrics, 74, 421-429.
}
\author{
.
}
