% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_functions.R
\name{LASSO}
\alias{LASSO}
\title{LASSO via glmnet, selecting a pre-specified number of features with non-zero coefficient}
\usage{
LASSO(df, time, status, features, dfmax = 100)
}
\arguments{
\item{df}{data frame including columns with features and time-to-event endpoint information}

\item{time}{character/column name containing survival time}

\item{status}{character/column name containing survival status}

\item{features}{vector of column names or column indices containing features}

\item{dfmax}{dfmax parameter in call to \code{glmnet}}
}
\value{
a vector with absolute coefficient for each feature
}
\description{
LASSO via glmnet, selecting a pre-specified number of features with non-zero coefficient
}
\details{
only a wrapper for \code{glmnet} function
}
\examples{
\dontrun{
require("survival")
require("penalized")
require("glmnet")
data(nki70)
LASSO(nki70,"time","event", colnames(nki70)[-c(1:7)])[1:10]
}
}
\references{
Simon, N., Friedman, J., Hastie, T., and Tibshirani, R. (2011). Regularization paths for Cox's proportional hazards model via coordinate descent. Journal of Statistical Software, 39, 1-13.
}
\author{
.
}
