% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_functions.R
\name{method.similarity}
\alias{method.similarity}
\title{method similarity between two methods}
\usage{
method.similarity(scores1, scores2, trueVars, sim.method = c("Jaccard",
  "corr"), m = 50)
}
\arguments{
\item{scores1}{scores matrix; output of screening methods for simulated data, larger value means better ranking, rows = simulation runs, columns = variables}

\item{scores2}{scores matrix; output of screening methods for simulated data, larger value means better ranking, rows = simulation runs, columns = variables}

\item{trueVars}{vector of indices of true variables}

\item{sim.method}{measure of similarity btw. methods; either Jaccard index for detected features or correlation of feature ranks}

\item{m}{model size for which to determine selected features; ignored if sim.method="corr"}
}
\value{
scalar/vector with FDR for each model size
}
\description{
method similarity between two methods
}
\details{
ties are handled using the 'random' approach in \code{rank}. average Pearson correlation btw. rankings of 2 methods for b'th simulation run. average Jaccard index for detected true variables at given model size
}
\author{
Manuela Hummel
}
