% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{methods.simplot}
\alias{methods.simplot}
\title{similarities btw. all methods, corrplot}
\usage{
methods.simplot(simoutput, trueVars, whichMethods, sim.method = c("Jaccard",
  "corr"), m = 50, order = c("hclust", "original", "AOE", "FPC",
  "alphabet"), hclust.method = c("complete", "ward", "ward.D", "ward.D2",
  "single", "average", "mcquitty", "median", "centroid"), ...)
}
\arguments{
\item{simoutput}{array w. output of simulation study from Axel}

\item{trueVars}{vector of indices of true variables}

\item{whichMethods}{names of methods that shall be included in the plot; if missing, all methods in simoutput are shown}

\item{sim.method}{measure of similarity btw. methods; either Jaccard index for detected features or correlation of feature ranks}

\item{m}{scalar or vector of model sizes}

\item{order}{if/how methods shall be ordered in corrplot, see ?corrplot, here default is "hclust"}

\item{hclust.method}{agglomeration method in case methods are ordered by 'hclust', see ?corrplot}

\item{...}{further parameters passed to corrplot()}
}
\description{
similarities btw. all methods, corrplot
}
\details{
wrapper for \code{corrplot} function R package corrplot
}
\author{
Manuela Hummel
}
