% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_functions.R
\name{mmms}
\alias{mmms}
\title{Median of Minimum Model Size needed to detect all true features (MMMS)}
\usage{
mmms(scores, trueVars, fractrue = 1)
}
\arguments{
\item{scores}{scores matrix; output of screening methods for simulated data, larger value means better ranking, larger value means better ranking}

\item{trueVars}{vector of indices of true variables}

\item{fractrue}{mmms for selecting \code{fractrue} * \code{trueVars} (e.g. 90\%) of the true variables}
}
\value{
scalar
}
\description{
Median of Minimum Model Size needed to detect all true features (MMMS)
}
\details{
ties are handled using the 'random' approach in \code{rank}
}
\author{
Manuela Hummel
}
