% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{perfplot}
\alias{perfplot}
\title{performance plot, visualization of performance vs model size}
\usage{
perfplot(scores, trueVars, what = c("TPR", "TDR", "FDR", "propallselected"),
  minm = 25, maxm, col = 1, lwd = 2, ylab, ylimit = c(0, 1),
  add = FALSE, ...)
}
\arguments{
\item{scores}{scores matrix; output of screening methods for simulated data, larger value means better ranking, larger value means better ranking}

\item{trueVars}{vector of indices of true variables}

\item{what}{performance criterion to be plotted (TPR, TDR, FDR, propallselected)}

\item{minm}{minimal model size}

\item{maxm}{maximal model size}

\item{col}{colour}

\item{lwd}{lwd}

\item{ylab}{ylab}

\item{ylimit}{limit of y-axis}

\item{add}{add to existing plot}

\item{...}{further parameters pased to plot function}
}
\description{
performance plot, visualization of performance vs model size
}
\details{
y = average TPR/TDR/FDR/propallselected, x = respective model size. vertical line indicates MMMS (median of minimum model size needed to detect all true variables). can be shown for all variables (m = 1,...,p) or up to 'maxm' variables (m = 1,...,maxm). show e.g. TPR and FDR for one method or compare e.g. TPR across methods
}
\author{
Manuela Hummel
}
