% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotall.perf}
\alias{plotall.perf}
\title{performance plot, visualization of performance vs model size for several methods}
\usage{
plotall.perf(files, trueVars, whichMethods, what = c("TPR", "TDR", "FDR",
  "propallselected"), ylimit = c(0, 1), maxm = 100)
}
\arguments{
\item{files}{paths + names of .RData files containing simulation results}

\item{trueVars}{vector of indices of true variables}

\item{whichMethods}{methods to be plotted}

\item{what}{performance criterion to be plotted (TPR, TDR, FDR, propallselected)}

\item{ylimit}{limit of y-axis}

\item{maxm}{maximal model size}
}
\description{
performance plot, visualization of performance vs model size for several methods
}
\details{
.
}
\author{
Manuela Hummel
}
