% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_functions.R
\name{propallselected}
\alias{propallselected}
\title{Proportion of simulation runs where all relevant variables are selected for given model size(s)}
\usage{
propallselected(scores, m, trueVars)
}
\arguments{
\item{scores}{scores matrix; output of screening methods for simulated data, larger value means better ranking, larger value means better ranking}

\item{m}{scalar or vector of model sizes}

\item{trueVars}{vector of indices of true variables}
}
\value{
scalar or vector with proportion of runs with all true variables being selected
}
\description{
Proportion of simulation runs where all relevant variables are selected for given model size(s)
}
\details{
ties are handled using the 'random' approach in \code{rank}
}
\author{
Manuela Hummel
}
