% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_functions.R
\name{propselected}
\alias{propselected}
\title{Proportion of simulation runs where single relevant variables are selected for given model size(s)}
\usage{
propselected(scores, m, trueVars)
}
\arguments{
\item{scores}{scores matrix; output of screening methods for simulated data, larger value means better ranking, larger value means better ranking}

\item{m}{scalar or vector of model sizes}

\item{trueVars}{vector of indices of true variables}
}
\value{
vector with selection proportions for true variables
}
\description{
Proportion of simulation runs where single relevant variables are selected for given model size(s)
}
\details{
ties are handled using the 'random' approach in \code{rank}
}
\author{
Manuela Hummel
}
