% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_functions.R
\name{RCDCS}
\alias{RCDCS}
\title{Robust Censored Distance Correlation Screening (RCDCS)}
\usage{
RCDCS(df, time, status, features)
}
\arguments{
\item{df}{data frame including columns with features and time-to-event endpoint information}

\item{time}{character/column name containing survival time}

\item{status}{character/column name containing survival status}

\item{features}{vector of column names or column indices containing features}
}
\value{
a named vector with squared distance correlation for each feature
}
\description{
Robust Censored Distance Correlation Screening (RCDCS)
}
\details{
.
}
\examples{
\dontrun{
require("survival")
require("Rfast")
require("penalized")
data(nki70)
RCDCS(nki70,"time","event", colnames(nki70)[-c(1:7)])[1:10]
}
}
\references{
Chen, X., Chen, X., and Wang, H. (2018). Robust feature screening for ultra-high dimensional right censored data via distance correlation. Computational Statistics & Data Analysis, 119,118-138.
}
\author{
Thomas Hielscher
}
