% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_functions.R
\name{RESI2}
\alias{RESI2}
\title{Distance correlation based on martingale residuals (RESI)}
\usage{
RESI2(df, time, status, features)
}
\arguments{
\item{df}{data frame including columns with features and time-to-event endpoint information}

\item{time}{character/column name containing survival time}

\item{status}{character/column name containing survival status}

\item{features}{vector of column names or column indices containing features}
}
\value{
a named vector of distance correlations
}
\description{
Distance correlation based on martingale residuals (RESI)
}
\details{
just a wrapper for the \code{distcorr} function
}
\examples{
\dontrun{
require("survival")
require("penalized")
data(nki70)
RESI2(nki70,"time","event", colnames(nki70)[-c(1:7)])[1:10]
}
}
\references{
.
}
\author{
Dominic Edelmann
}
