% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_functions.R
\name{SIS}
\alias{SIS}
\title{SIS-Cox using absolute Pearson correlation of feature values and survival time}
\usage{
SIS(df, time, status, features)
}
\arguments{
\item{df}{data frame including columns with features and time-to-event endpoint information}

\item{time}{character/column name containing survival time}

\item{status}{character/column name containing survival status}

\item{features}{vector of column names or column indices containing features}
}
\value{
a named vector with absolute correlation coefficient for each feature
}
\description{
SIS-Cox using absolute Pearson correlation of feature values and survival time
}
\details{
Ignores survival status
}
\examples{
\dontrun{
require("survival")
require("penalized")
data(nki70)
SIS(nki70,"time","event", colnames(nki70)[-c(1:7)])[1:10]
}
}
\references{
Saldana, D. F. and Feng, Y. (2018). SIS: An R package for sure independence screening in ultrahigh-dimensional statistical models. Journal of Statistical Software, 83, 1-25.
}
\author{
.
}
