% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_functions.R
\name{TPR}
\alias{TPR}
\title{True Positive Rate (TPR) for given model size(s)}
\usage{
TPR(scores, m, trueVars)
}
\arguments{
\item{scores}{scores matrix; output of screening methods for simulated data, larger value means better ranking, larger value means better ranking}

\item{m}{scalar or vector of model sizes}

\item{trueVars}{vector of indices of true variables}
}
\value{
scalar/vector with TPR for each model size
}
\description{
True Positive Rate (TPR) for given model size(s)
}
\details{
ties are handled using the 'random' approach in \code{rank}. the average number of true variables contained in the top m selected we get by multiplying TPR by the number of true variables
}
\author{
Manuela Hummel
}
